--- console/src/components/Common/GraphQLEditor/GraphQLEditor.scss
@@ -1,5 +1,0 @@
-@import '../Common.scss';
-
-.errorMessage {
-  color: red;
-}
